PersonCard: display tags in sorted order
The code in PersonCard for displaying Tags as Labels is as follows:

    person.getTags().forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));

However, `person.getTags()` is a Set (stored internally as a HashSet).
Sets do not provide any ordering guarantees, which means that the order
in which tags are displayed in the PersonCard is potentially
non-deterministic. This means that two people having the same tags could
potentially have their tags displayed in different orders, which could
confuse and annoy users who expect AddressBook to behave in a
predictable manner.

Fix this by sorting the tags first before adding them as Labels, thus
imposing an ordering on the tags. Lexicographical ordering based on the
tag name is used since that's a commonly used ordering criterion.