Model: refactor package `person` to `module`
We are going to morph a `persons`-oriented Address Book application to a
`modules`-oriented Degree Planner application.

Before we can start tweaking existing features and implementing new
features to suit our degree planner application, we need to refactor the
codebase so that we are working with `modules` instead of `persons`.

As the codebase is quite huge, we carry out the refactoring in multiple
stages incrementally to ensure the correctness of the refactoring done.

In the first phase of refactoring, we retain all existing
functionalities, but avoid updating any variables (such as strings or
variable names referenced by Jackson JSON processor for serialization)
or filenames that are referenced by unit tests so as to allow all unit
tests to continue to pass.

In the second phase of refactoring, we update all variables and
filenames, and amend unit tests accordingly so that all unit tests
continue to pass.

In the third phase of refactoring, we update the documentations
(e.g. User Guide and Developer Guide).

As an initial step of the first refactoring phase, let's rename the
model package `person` to `module`.