UI: refactor class `PersonListPanel` to `ModuleListPanel`
We are in the first refactoring phase of morphing `persons`-oriented
Address Book application to a `modules`-oriented Degree Planner
application.

In this first phase of refactoring, let's refactor class
`PersonListPanel` to `ModuleListPanel` and:
* replace `*person*` with `*module*` in comments
* cascade the changes to all affected classes

Note: ListView<Module> `personListView` remains unchanged in this commit
as it will affect how JavaFX renders the FXML, which might cause unit
tests to fail. As such, this class attribute `personListView` will only
be refactored in the second phase of refactoring.