Storage: refactor `JsonSerializableAddressBook#persons` to `modules`
We are in the second refactoring phase of morphing `persons`-oriented
Address Book application to a `modules`-oriented Degree Planner
application.

In this second phase of refactoring, let's refactor
`JsonSerializableAddressBook#persons` class attribute to `modules` and
update all test files in `JsonAddressBookStorageTest` directory to
reflect the change in naming of the Jackson JsonProperty.