docs: replace `*person*` with `*module*` in User and Developer guides
We are almost done with the third and final refactoring phase of
morphing `persons`-oriented Address Book application to a
`modules`-oriented Degree Planner application.

As a final step of this third refactoring phase, let's update the User
images shown in the Developer Guide to replace all instances of
`*person*` with `*module*`.