Storage: implementing a new storage model
After a series of discussion and considerations, the group have
decided to implement a new storage model. Where 'JsonAddressBookStorage'
will now only be able to interact with 'JsonSerializableAddressBook'.

'JsonSerialiazbleAddressBook' will then interact with the following 3
classes:
- 'JsonSerializableRequirementCategoryList'
- 'JsonSerializableModuleList'
- 'JsonSerializableDegreePlannerList'

'JsonAddressBookStorage' will not be able to interact with the above 3
classes. It will call methods in 'JsonSerializableAddressBook' instead.

The following are change made in this commit:

- 'RequirementCategory' will now have its own data storage file as
  defined in'UserPref'. Previously it was using the same file as
  addressBook

- 'JsonSeriliazbleApplication' is created with the relevant logic to
  read and write to data storage files

- 'JsonSerializableRequirementCategoryList' is created with the relevant
  methods to define how data is stored

- 'JsonAddressBookStorage' is updated accordingly to the above changes