Storage: update logic flow of reading and storing from Json files
The current implementation now is where the
'JsonSerializableApplication' will handle the read and write operations
to the existing data storage files.

This should not be the case as 'JsonAddressBookStorage' should be the
one handling these operations. While 'JsonSerializableApplication'
should only be interacting with the classes of:
'JsonSerialiazbleAddressBook'
'JsonSerializableRequirementCategoryList'
'JsonSerializableDegreePlannerList'

The following are changes made in this commit:

- Move reading and saving of storage data files from
  'JsonSerializableApplication' to 'JsonAddressBookStorage'

- Initialize methods in 'JsonSerializableApplication' to return the
  classes of the above mentioned files

- Update to 'LogicManager', 'Storage', 'StorageManager' in accordance with
  the above changes