RequirementCategoryBuilder: update to faciliate new command
As a new command will be added, the 'RequirementCategoryBuilder' was
validated to see if there are any conflicting attributes or methods
that may cause a problem at a later stage.

Upon insepction, a few comments and methods have to be updated. Let's
update the 'RequirementCategoryBuilder' class so that it will not have
any conflicting issues when the new command is implemented.