commands: add 'RequirementAddCommand' class with uninitialized methods
In order for the program to be able to handle a new command, a class
must first be created to handle the logic flow for the new command.

In this case, a command to add a module to an existing requirment
category is to be implemented.

Let's create a skeleton 'RequirmentAddCommand' class with uninitialzed
methods to show how the logic would flow for the new command. Let's also
add in instructions on how the command is to be used as well as error
messages that will be displayed if an error was to occur.