commands: initialize methods created in 'RequirementAddCommand' class
The methods created in the 'RequirmentAddCommand' class is to have a
skeleton for the logic flow of the new command. However the methods
created have not yet been referenced in the 'Model', 'ModelManager',
'AddressBook' and 'AddCommandTest'

Let's add the method created in the 'RequirementAddCommand' class so
that the application is able to reference to these methods.