UniqueRequirementCategoryList: implement logic of initialized methods
The skeleton for the new command have been created where a module is
added to an existing requirment category in the 'RequirementAddCommand'
class. However, the logic to how the commands work have not been added
yet.

Let's add the logic for the commands in the
'UniqueRequirementCategoryList' class to allow the new command to work
properly.

* The 'location()' method will return the position of the desired
requirement category in the requirment category list

* The 'isModuleInRequirmentCategory()' method checks if the requirement
category already contains the module to be added

* The 'doesModuleExistInApplication()' method checks if a module exists in the
existing moduleList in the application

* The 'addModuleToRequirmentCategory()' method adds the module entered
to a requirement category