parser: implementing 'RequirementAddCommandParser' for the new command
After the skeleton and logic for the new command is implemented, the
way the data is read from the user's input must be implemented as well
so that the data can be passed to the logic and then stored into the
data storage files.

Let's implement the parser needed in order to read the user's input data
and pass it to 'RequirmentAddCommand' class.

* 'RequirmentAddCommandParser' class was created to parse the input data
based on the 'name' and 'code' prefixes.

- The 'RequirmentAddCommandParser' parsers the name to be the requirment
  category to add the module to

- The 'RequirementAddCommandParser' parsers the code to be the modules
  that will be added to the requirment category

- The 'RequirementAddCommandParser' will then create a temporary
  list of modules with temporary data, except for the module codes
  which will be validated in the 'RequirementAddCommand' class

* A 'parseCodes' method was created in 'ParserUtil' to return back a
  list of module codes