View: update MainWindow to include UI for DegreePlannerList
Currently, MainWindow does not include UI for DegreePlannerList.
Moreover, BrowserPanel is not needed to be displayed.

Let's update MainWindow to display Ui for DegreePlannerList and minimize
the size for BrowserPanel and place it at the bottom so that it does not
interrupt other pannels.

In addition, with UI for DegreePlannerList included, it will have problem
activating 'Help' button on the menu when MainWindow is resized to
current minimum size.

Hence, let's increase the current minimum size for MainWindow to prevent
this problem.