Model: update `Module#isSameModule(...)` (#85)
When comparing if two modules are same using `Module#isSameModule(...)`,
the equality of `Name`, `Code` and `Credits` fields are compared.

However, this comparison may be too strict, and may result in duplicate
modules being created.

e.g. `Module#isSameModule(...)` will return false for the below modules:

  Code: AA1234
  Name: Intro to ABC
  Credits: 4

  Code: AA1234
  Name: Intro to ABC
  Credits: 12

However, they are logically referring to the same module.

We should instead make `Code` the identity field, and make `Name`,
`Credits` and `Tag` data fields of `Module` objects.

Firstly, `Credits` or `Tag` should not be considered as identity fields
because their values are not unique enough.

Secondly, `Name` should not be considered as an identity field because
some modules have identical names, but refer to different modules.

For instance, the below two modules are two separate entities:

  Code: GEK1531
  Name: Cyber Security
  Credits: 4

  Code: GET1004
  Name: Cyber Security
  Credits: 4

Let's update the `Module#isSameModule(...)` equality comparison to
reflect the changes in identity fields for `Module` objects, and fix the
affected unit tests as well.