requirement: update logic of new command and checks
The current implementations, creates a new RequirementCategory object
with the parsed data and a temp Credit variable. The RequirementCategory
objected created will then be passed to the 'RequirementAddCommand'
class.

The new logic update will not create a new RequirementCategory object
and instead only pass data that was parsed. In this case, only the Name
and Set<Code> will be passed to 'RequirementAddCommand' class.

The current implementation of the 'requirement_add' command has a bug
where no more than 1 module code can be added to a requirement category.
The bug was caused by not using a new Set<Code> variable, when a module
code is added to a requirement category with a module code already added
to it.

The update of the logic when a module code is added to a requirement
category now adds to a new Set<Code> variable and updates the total
amount of credits of a requirement category based on the modules added
to the requirement category.

The location method was updated to return an RequirementCategory object
instead of a interger value.

The other methods were updated accordingly to the data that was parsed,
instead of a 'RequirementCategory' object.