RequirementAddCommand: update methods and to receive data parsed
The current 'RequirementAddCommand' class was to receive a
RequirementCategory object. The RequirementCategory object will then
passed to various methods for checking and validation before adding
the module code to the requirement category.

The updated 'RequirementAddCommand' class now only takes in the data
that was parsed only and then pass the relevant data to methods for
checking and validation.

The name of the error messages are also updated according to the methods
name.