Logic: create `BooleanExpressionParser` class
Our implementation for `FindCommandParser#getKeywordsPredicate` result
an implicit boolean OR predicate.

This is not desirable for our users as they will not be able to have a
tighter search results. e.g. Finding for module's `name` with both
`Programming` and `Methodology`.

With the aim to improve our `FindCommand`

let's create a new class skeleton `BooleanExpressionParser` to help
us evaluate the argument which the user provide.