Logic: create `Operator` class
Our implementation for `FindCommandParser#getKeywordsPredicate` result
an implicit boolean OR predicate.

This is not desirable for our users as they will not be able to have a
tighter search results. e.g. Finding for module's `name` with both
`Programming` and `Methodology`.

Let's create an enum class on all the operators that will be used by
`BooleanExpressionParser` and include:
- precedence to indicate which operators should be evaluated first.