Logic: implement `BooleanExpressionParser#parse`
Our implementation for `FindCommandParser#getKeywordsPredicate` result
an implicit boolean OR predicate.

This is not desirable for our users as they will not be able to have a
tighter search results. e.g. Finding for module's `name` with both
`Programming` and `Methodology`.

Let's improve our parsing method to make use of `Shunting Yard` so we
could detect if `&&` or `||` is used and parse according to what the
user provide.