Model/util: add and use `InitialDataUtil` class (#134)
* Model/util: add `InitialDataUtil` class

Whenever the `clear` command is executed, or when the application fails
to read the JSON save data successfully, `new AddressBook()` is passed
to `ModelManager#setAddressBook(...)` to effectively 'clear' all save
data.

However, this behaviour is undesirable, as it purges all existing
modules, requirement categories and degree planners. This causes users
to be unable to interact with requirement categories and degree planners
as they no longer exists in the application.

As such, instead of purging all the data by replacing the current
`AddressBook` to a new instance, we should replace it to a barebone copy
of an empty AddressBook that retains the respective requirement
categories and degree planner.

Let's add an `InitialDataUtil` class to provide easy access of reseting
existing data to the initial (empty) data described above.

* Replace `new AddressBook()` with `getInitialAddressBook()`

Let's replace `new AddressBook()` with `getInitialAddressBook()`
whenever `clear` command is executed, or when the application fails
to read the JSON save data successfully, and update the unit tests for
the `clear` command as well.