Logic/parser: add and use `BooleanExpressionTokenizer`
`BooleanExpressionParser` uses `StringTokenizer` to tokenize the command
arguments.

However, `StringTokenizer` is insufficient in extracting prefixed
arguments and boolean operators, as we want to keep boolean operators as
tokens. Likewise, `ArgumentTokenizer` is also not suitable, as we do not
want to use whitespace as delimiter for splitting and extracting
arguments.

Consider the following example.

  find name/Structures || (code/CS2040C && credits/4)

Both `StringTokenizer` and `ArgumentTokenizer` are unable to split the
tokens into:
* name/Structures
* ||
* (
* code/CS2040C
* &&
* credits/4
* )

But, we want to be able to extract the tokens as above to enable
Shunting-yard algorithm to work nicely without any additional processing
of arguments in `BooleanExpressionParser`.

To enable this, let's create `BooleanExpressionTokenizer`, a string
tokenizer class that extracts all arguments with the specified prefixes
and boolean operators as tokens, and replace the use of
`StringTokenizer` to `BooleanExpressionTokenizer` in
`BooleanExpressionParser`.