Logic: Improve `BooleanExpressionParser#getKeywordsPredicate`
With the implementation of Shunting Yard algorithm, we do not need to
parse multiple keywords into its predicate.

Let's improve on our creation of predicate by doing the following:

If the search term is just a single word.
We will find if the keyword is contain within its particular attributes.

If the search term is a sentence.
We wll find if the keyword matches the sentence exactly.

and:
- add new unit tests for method added.
- remove unused method.