Logic: create logic for 'requirement_move' command
Currently our application does not allow users to move the modules that
have been added to requirement categories.

This makes it difficult for users when they are moving the modules in
the requirement category around. They would have to delete and add the
modules multiple times during the moving process.

To address these issues, let's add the 'requirement_move' command to
allow users to move modules in the requirement category.

First, let's implement the logic to handle how the modules are going to
be moved from one requirement category to another.

- 'RequirementMoveCommandParser' is used to parse in the user input

- 'RequirementMoveCommand' is used to handle the parsed input and this
 class is where most of the logic reside