Logic/parser: convert `BooleanExpressionParser` to use generics (#193)
The `BooleanExpressionParser` is constrained to only finding `Module` 
only. However, this means that we are unable to extend the 
`BooleanExpressionParser` to find other objects, such as 
`DegreePlanner`.

Let's convert `BooleanExpressionParser` to use generics so as to allow
finding of other objects based on boolean expressions and return the 
custom `Predicate`.