Model/module: add `Module#semesters`
Users may want to keep track of what modules are offered in an academic
year.

For instance, CS1010 Programming methodology is offered in Semesters
1, 2, 4 (Special Semester Term 2) only. It would be pointless for users
to plan to take the module in Special Semester Term 1 (Semester 3 in our
application) when it is not actually being offered.

Let's implement `Module#semesters` to enable users to keep track of
which semesters a `Module` can be taken, and update the UI to show the
semesters the module is offered in.