Logic: update 'planner_move' to handle co-requisites
The application allows the user to set co-requisites to modules.
However, 'planner_move' does not handle co-requisites.
Hence, the user has to move the codes that shares co-requisites
manually.

Let's enhance 'planner_move' command by allowing it to handle
co-requisites so that when a module is moved, its co-requisites are
moved along if they exist in the degree plan.