Logic: Refactor 'planner_list_all' to 'planner_list'
After the discussion, we decided to refactor original 'planner_list'
to 'planner_show'.

In order to be consistent with 'list' command, let's refactor the
current 'planner_list_all' command to 'planner_list'.

This consistency will allow a user to easily perceive the command
usage and less likely to cause the confusion when trying to figure
out the command for listing the degree plan.

In addition, let's update 'planner_list' to utilize
 'DegreePlanner:toString' method instead of current method using
'StringBuilder' since the format for 'DegreePlanner:toString' was
refined to print each 'DegreePlanner' nicely.