DeveloperGuide: update "Troubleshooting Testing" section (#219)
Due to new privacy enhancements introduced in macOS Mojave, applications
without `Accessibility` permission cannot simulate keyboard and mouse
movements.

This causes GUI tests that uses `CommandBoxHandle#run(...)` to execute
commands to fail, as `CommandBoxHandle#run(...)` attempts to simulate
pressing `Enter` key in the Command Box using `GuiRobot` (subclass of
`org.testfx.api.FxRobot`).

The `Enter` keypress event is then handled by
`CommandBox#handleCommandEntered()`, which executes the command.

However, because of insufficient privileges, `GuiRobot` may be unable to
simulate the `Enter` keypress event in the Command Box, thus the handler
method is not called. This results in the command in the Command Box to
not get executed, which in turn causes the GUI tests to fail.

To reduce confusion amongst developers using macOS Mojave, let's update
the "Troubleshooting Testing" section to discuss this issue, and provide
instructions on how to resolve it.