[#191] Logic/commands: implement `planner_suggest` command (#192)
Users may not know what module(s) to put in the degree plan. 

Let's implement `planner_suggest` to recommend module(s) users can add 
to the degree plan.

* `planner_suggest` requires users to input desirable credits and tags, 
and modules with the highest number of matching tags and does not exist 
in the degree plan will be suggested first. 
* The modules with closest credits will be suggested if there is a tie. 
* Tag-matching is prioritized in the recommendation; if there is no tag 
matched, no module will be shown.
* `tag` is optional parameter. If no `tag` is in input, the 
recommendation will be purely based on credits.
* Modules with matching tags and credits that are not in the degree plan
are given as supplementary information.
* User Guide will be updated at a later time.